refactor: rename v2.0.0 upgrade to v1.6.0
