+++ client/app/upgrades/v_1_6_0/upgrades_test.go
@@ -1,4 +1,4 @@
-package v_2_0_0
+package v_1_6_0
 
 import (
 	"context"
